% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_package}
\alias{init_package}
\title{Checks and sets environmental variables}
\usage{
init_package(R_DATA_PATH = NULL,
  renviron_path = Sys.getenv("R_ENVIRON_USER"))
}
\arguments{
\item{renviron_path}{Path to .Renviron file}
}
\description{
The package relies on several environmental variables. On load the package tests whether the variables are set, if not it prompts for setting them. The function `init_package` can be also called within the session to overwrite the current setting.
}
\details{
The function simply writes the environmental variables into .Renviron file (typically in the home directory). If you encounter problems with calling the function or during package load, try to alter .Renviron file manually.

The contents of the .Renviron can be e.g.

R_DATA_PATH='/home/DATA'

Currently implemented are:

\describe{
  \item{`R_DATA_PATH`}{The path to store big data. Only local storage implemented at the moment.}
}
}
