% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{rotated_grid_transform}
\alias{rotated_grid_transform}
\title{Transforms coordinates to rotated grid}
\usage{
rotated_grid_transform(grid_in, option = "2rot", SP_coor = c(39.25,
  -162))
}
\arguments{
\item{grid_in}{matrix of lon-lat coordinates}

\item{option}{either `2rot` - transforms to rotated grid, other values transform from rotated grid to lon-lat}

\item{SP_coor}{a vector of `grid_north_pole_latitude` and `grid_north_pole_longitude` from the netcdf file.}
}
\value{
data.table with coordinates
}
\description{
Transforms coordinates to rotated grid
}
