% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{rotatePoints}
\alias{rotatePoints}
\title{Rotate spatial points}
\usage{
rotatePoints(shp, ...)
}
\arguments{
\item{shp}{SpatialPoints or SpatialPointsDataFrame}

\item{...}{}
}
\value{
SpatialPoints at rotated coordinates
}
\description{
The functions uses `rcm2019::rotated_grid_transform` to rotate coordinates of the points from the shape file
}
