% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{rotatePoly}
\alias{rotatePoly}
\title{Transforms shapefile to rotated coordinates}
\usage{
rotatePoly(shp, ...)
}
\arguments{
\item{shp}{SpatialPolygons or SpatialPolygonsDataFrame}
}
\value{
SpatialPolygons at rotated coordinates
}
\description{
The functions uses `rcm2019::rotated_grid_transform` to rotate coordinates of the polygons from the shape file
}
