% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setdp}
\alias{setdp}
\title{Sets working directory on the R_DATA_PATH}
\usage{
setdp(dir = "")
}
\arguments{
\item{dir}{subdirectory of R_DATA_PATH}
}
\description{
Simply wrapper around `setwd(file.path(Sys.getenv('R_DATA_PATH'), dir))`
}
